+++ lib/adk_core.dart
@@ -50,6 +50,7 @@ export 'src/models/llm_request.dart';
 export 'src/models/llm_response.dart';
 export 'src/models/base_llm.dart';
 export 'src/models/google_llm.dart';
+export 'src/models/lite_llm.dart';
 
 export 'src/plugins/base_plugin.dart';
 export 'src/plugins/plugin_manager.dart';